[COUR-2277] added shortcut to return default DateTime if empty string is passed to DateHelper
